Handle attachments with CHILD_PART_CONTAINS_DATA data location

+++ k9mail-library/src/main/java/com/fsck/k9/mail/Message.java
@@ -145,7 +145,7 @@ public abstract class Message implements Part, Body {
 
     public abstract boolean hasAttachments();
 
-    public abstract int getSize();
+    public abstract long getSize();
 
     public void delete(String trashFolderName) throws MessagingException {}
