+++ k9mail-library/src/main/java/com/fsck/k9/mail/internet/MimeMessage.java
@@ -182,7 +182,7 @@ public class MimeMessage extends Message {
     }
 
     @Override
-    public int getSize() {
+    public long getSize() {
         return mSize;
     }
